vision-engine/requirements.txt
# Meowdel Real-Time Vision Processing Engine
# Python + Flask + ML for continuous video analysis

# Web Framework
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.11.1
python-engineio==4.9.0
eventlet==0.35.2

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# Machine Learning
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.4.1.post1
numpy==1.26.4
scipy==1.12.0

# Facial Analysis
face-recognition==1.3.0
dlib==19.24.2
deepface==0.0.89

# OCR (for document reading)
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.3

# Image Processing
Pillow==10.2.0
imageio==2.34.0

# Text Processing
anthropic==0.18.1  # For cat personality
openai==1.12.0     # Fallback

# Utilities
python-dotenv==1.0.1
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
aiohttp==3.9.3

# Monitoring
prometheus-client==0.20.0
sentry-sdk==1.40.6

# Development
pytest==8.0.2
black==24.2.0
flake8==7.0.0
